pdf_to_docx/requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
